UPLUS-126 RDB 기반 사용량 배치 시스템 보완

+++ src/main/java/com/project/rdb/batch/model/entity/UsageNotificationOutbox.java
@@ -38,7 +38,7 @@ public class UsageNotificationOutbox {
     private int threshold;
 
     @Column(nullable = false)
-    private int percent;
+    private double percent;
 
     @Column(name = "total_used_mb", nullable = false)
     private Long totalUsedMb;
